Swap alt and ⌘ (suplm) to match corresponding swap of keys on keyboard
This makes alt, not suplm (formerly referred to as “win”), be the window-manipulation modifier key, and suplm, not alt, be the secondary-function modifier key. With the corresponding swap of mappings on the keyboard, this keeps the window-manipulation and secondary-function modifier keys unmoved physically. The purpose is to effect physical movement of AltGr, which is the same physical key as suplm (latter being accessed via fn layer).

First change of vimizer.el in almost 5 years (last time was 2015 Sept 12), and the change is only for the sake of this swap of key bindings.